+++ src/main/java/com/github/attacktive/msaproduct/product/adapter/inbound/ProductService.java
@@ -1,13 +1,14 @@
-package com.github.attacktive.msaproduct.product.service;
+package com.github.attacktive.msaproduct.product.adapter.inbound;
 
 import java.util.List;
 import java.util.Set;
 
-import com.github.attacktive.msaproduct.product.api.exception.NoSuchProductException;
-import com.github.attacktive.msaproduct.product.api.request.AddProductRequest;
-import com.github.attacktive.msaproduct.product.api.request.UpdateProductRequest;
+import com.github.attacktive.msaproduct.product.application.exception.NoSuchProductException;
+import com.github.attacktive.msaproduct.product.application.request.AddProductRequest;
+import com.github.attacktive.msaproduct.product.application.request.UpdateProductRequest;
 import com.github.attacktive.msaproduct.product.domain.Product;
-import com.github.attacktive.msaproduct.product.repository.ProductRepository;
+import com.github.attacktive.msaproduct.product.port.inbound.ProductUseCase;
+import com.github.attacktive.msaproduct.product.adapter.outbound.ProductRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.lang.Nullable;
