+++ src/main/java/com/github/attacktive/msaproduct/product/application/ProductController.java
@@ -1,13 +1,13 @@
-package com.github.attacktive.msaproduct.product.api;
+package com.github.attacktive.msaproduct.product.application;
 
 import java.util.List;
 import java.util.Set;
 import javax.validation.Valid;
 
-import com.github.attacktive.msaproduct.product.api.request.AddProductRequest;
-import com.github.attacktive.msaproduct.product.api.request.UpdateProductRequest;
+import com.github.attacktive.msaproduct.product.application.request.AddProductRequest;
+import com.github.attacktive.msaproduct.product.application.request.UpdateProductRequest;
 import com.github.attacktive.msaproduct.product.domain.Product;
-import com.github.attacktive.msaproduct.product.service.ProductUseCase;
+import com.github.attacktive.msaproduct.product.port.inbound.ProductUseCase;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
