+++ src/main/java/com/github/attacktive/msaproduct/product/application/configuration/WebMvcConfiguration.java
@@ -1,4 +1,4 @@
-package com.github.attacktive.msaproduct.product.api.configuration;
+package com.github.attacktive.msaproduct.product.application.configuration;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Configuration;
