+++ src/main/java/com/github/attacktive/msaproduct/product/application/response/ErrorResponse.java
@@ -1,4 +1,4 @@
-package com.github.attacktive.msaproduct.product.api.response;
+package com.github.attacktive.msaproduct.product.application.response;
 
 import com.fasterxml.jackson.annotation.JsonGetter;
 import com.fasterxml.jackson.annotation.JsonIgnore;
