+++ src/main/java/com/github/attacktive/msaproduct/product/port/inbound/ProductUseCase.java
@@ -1,10 +1,10 @@
-package com.github.attacktive.msaproduct.product.service;
+package com.github.attacktive.msaproduct.product.port.inbound;
 
 import java.util.List;
 import java.util.Set;
 
-import com.github.attacktive.msaproduct.product.api.request.AddProductRequest;
-import com.github.attacktive.msaproduct.product.api.request.UpdateProductRequest;
+import com.github.attacktive.msaproduct.product.application.request.AddProductRequest;
+import com.github.attacktive.msaproduct.product.application.request.UpdateProductRequest;
 import com.github.attacktive.msaproduct.product.domain.Product;
 import org.springframework.lang.Nullable;
