+++ src/main/java/com/github/attacktive/msaproduct/product/port/outbound/persristence/ProductEntity.java
@@ -1,4 +1,4 @@
-package com.github.attacktive.msaproduct.product.repository.persistence;
+package com.github.attacktive.msaproduct.product.port.outbound.persristence;
 
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
@@ -6,7 +6,7 @@ import javax.persistence.GenerationType;
 import javax.persistence.Id;
 import javax.persistence.Table;
 
-import com.github.attacktive.msaproduct.product.api.request.ProductRequest;
+import com.github.attacktive.msaproduct.product.application.request.ProductRequest;
 import com.github.attacktive.msaproduct.product.domain.Product;
 import lombok.NoArgsConstructor;
